Added Auth for userDelete

+++ src/main/java/dhbw/mwi/Auslandsemesterportal2016/servlets/UserDeleteServlet.java
@@ -13,17 +13,24 @@ public class UserDeleteServlet extends HttpServlet {
 
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws IOException {
-        String matrikelnummer = request.getParameter("matrikelnummer");
-        PrintWriter toClient = response.getWriter();
+      int rolle = userAuthentification.isUserAuthentifiedByCookie(request);
 
-        if (matrikelnummer != null) {
-                String query = "DELETE FROM user WHERE matrikelnummer = ?";
-                String[] args = new String[]{matrikelnummer};
-                String[] types = new String[]{"int"};
-                SQL_queries.executeUpdate(query,args,types);
-        } else {
-            response.setStatus(HttpServletResponse.SC_INTERNAL_SERVER_ERROR);
-            toClient.println("Error: parameter are missing");
-        }
+      if(rolle!=1 && rolle!=2){
+        response.sendError(401,"Rolle: "+rolle);
+      }
+      else{      
+          String matrikelnummer = request.getParameter("matrikelnummer");
+          PrintWriter toClient = response.getWriter();
+
+          if (matrikelnummer != null) {
+                  String query = "DELETE FROM user WHERE matrikelnummer = ?";
+                  String[] args = new String[]{matrikelnummer};
+                  String[] types = new String[]{"int"};
+                  SQL_queries.executeUpdate(query,args,types);
+          } else {
+              response.setStatus(HttpServletResponse.SC_INTERNAL_SERVER_ERROR);
+              toClient.println("Error: parameter are missing");
+          }
+      }
     }
 }
